requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandasql>=0.7.3